+++ src/error.rs
@@ -48,20 +48,4 @@ impl fmt::Display for ParseError {
     }
 }
 
-/// An error when appending to a set.
-#[derive(Debug, Clone, Copy, PartialEq, Eq)]
-pub enum InsertError {
-    /// The set is full.
-    Full,
-    /// The item is already in the set.
-    Duplicate,
-}
-
-impl fmt::Display for InsertError {
-    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-        match self {
-            InsertError::Full => write!(f, "Set is full."),
-            InsertError::Duplicate => write!(f, "Item already exists."),
-        }
-    }
-}
+impl std::error::Error for ParseError {}